+++ rust/operator-binary/src/command.rs
@@ -9,13 +9,14 @@ use stackable_operator::{
 use crate::{
     authentication::TrinoAuthenticationConfig,
     catalog::config::CatalogConfig,
+    config::{client_protocol, fault_tolerant_execution},
     controller::{STACKABLE_LOG_CONFIG_DIR, STACKABLE_LOG_DIR},
     crd::{
-        CONFIG_DIR_NAME, Container, LOG_PROPERTIES, RW_CONFIG_DIR_NAME, STACKABLE_CLIENT_TLS_DIR,
+        CONFIG_DIR_NAME, Container, EXCHANGE_MANAGER_PROPERTIES, LOG_PROPERTIES,
+        RW_CONFIG_DIR_NAME, SPOOLING_MANAGER_PROPERTIES, STACKABLE_CLIENT_TLS_DIR,
         STACKABLE_INTERNAL_TLS_DIR, STACKABLE_MOUNT_INTERNAL_TLS_DIR,
         STACKABLE_MOUNT_SERVER_TLS_DIR, STACKABLE_SERVER_TLS_DIR, STACKABLE_TLS_STORE_PASSWORD,
-        SYSTEM_TRUST_STORE, SYSTEM_TRUST_STORE_PASSWORD, TrinoRole,
-        fault_tolerant_execution::ResolvedFaultTolerantExecutionConfig, v1alpha1,
+        SYSTEM_TRUST_STORE, SYSTEM_TRUST_STORE_PASSWORD, TrinoRole, v1alpha1,
     },
 };
 
@@ -23,7 +24,8 @@ pub fn container_prepare_args(
     trino: &v1alpha1::TrinoCluster,
     catalogs: &[CatalogConfig],
     merged_config: &v1alpha1::TrinoConfig,
-    resolved_fte_config: &Option<ResolvedFaultTolerantExecutionConfig>,
+    resolved_fte_config: &Option<fault_tolerant_execution::ResolvedFaultTolerantExecutionConfig>,
+    resolved_spooling_config: &Option<client_protocol::ResolvedClientProtocolConfig>,
 ) -> Vec<String> {
     let mut args = vec![];
 
@@ -85,13 +87,17 @@ pub fn container_prepare_args(
         args.extend_from_slice(&resolved_fte.init_container_extra_start_commands);
     }
 
+    // Add the commands that are needed for the client spooling protocol (e.g., TLS certificates for S3)
+    if let Some(resolved_spooling) = resolved_spooling_config {
+        args.extend_from_slice(&resolved_spooling.init_container_extra_start_commands);
+    }
+
     args
 }
 
 pub fn container_trino_args(
     authentication_config: &TrinoAuthenticationConfig,
     catalogs: &[CatalogConfig],
-    resolved_fte_config: &Option<ResolvedFaultTolerantExecutionConfig>,
 ) -> Vec<String> {
     let mut args = vec![
         // copy config files to a writeable empty folder
@@ -119,12 +125,17 @@ pub fn container_trino_args(
         }
     });
 
-    // Add fault tolerant execution environment variables from files
-    if let Some(resolved_fte) = resolved_fte_config {
-        for (env_name, file) in &resolved_fte.load_env_from_files {
-            args.push(format!("export {env_name}=\"$(cat {file})\""));
-        }
-    }
+    // Resolve credentials for fault tolerant execution exchange manager if needed
+    args.push(format!(
+        "test -f {rw_exchange_manager_config_file} && config-utils template {rw_exchange_manager_config_file}",
+        rw_exchange_manager_config_file = format!("{RW_CONFIG_DIR_NAME}/{EXCHANGE_MANAGER_PROPERTIES}")
+    ));
+
+    // Resolve credentials for spooling manager if needed
+    args.push(format!(
+        "test -f {rw_spooling_config_file} && config-utils template {rw_spooling_config_file}",
+        rw_spooling_config_file = format!("{RW_CONFIG_DIR_NAME}/{SPOOLING_MANAGER_PROPERTIES}")
+    ));
 
     args.push("set -x".to_string());