+++ rust/operator-binary/src/config/mod.rs
@@ -1,1 +1,4 @@
+pub mod client_protocol;
+pub mod fault_tolerant_execution;
 pub mod jvm;
+pub mod s3;