+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Support for fault-tolerant execution ([#779]).
+- Support for fault-tolerant execution ([#779], [#793]).
+- Support for the client spooling protocol ([#793]).
 - Helm: Allow Pod `priorityClassName` to be configured ([#798]).
 
 [#779]: https://github.com/stackabletech/trino-operator/pull/779
+[#793]: https://github.com/stackabletech/trino-operator/pull/793
 [#798]: https://github.com/stackabletech/trino-operator/pull/798
 
 ## [25.7.0] - 2025-07-23